Move duplicate event listeners to utils funciton
